+++ src/Csp.php
@@ -3,13 +3,11 @@ declare(strict_types = 1);
 
 namespace Middlewares;
 
-use Middlewares\Utils\Factory;
 use ParagonIE\CSPBuilder\CSPBuilder;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\MiddlewareInterface;
 use Psr\Http\Server\RequestHandlerInterface;
-use Psr\Log\LoggerInterface;
 
 class Csp implements MiddlewareInterface
 {
@@ -17,16 +15,6 @@ class Csp implements MiddlewareInterface
      * @var CSPBuilder|null
      */
     private $builder;
-
-    /**
-     * @var string|null
-     */
-    private $reportPath = null;
-
-    /**
-     * @var LoggerInterface|null
-     */
-    private $reportLogger = null;
 
     public static function createFromFile(string $path): self
     {
@@ -47,26 +35,10 @@ class Csp implements MiddlewareInterface
     }
 
     /**
-     * Configure the report-uri directive
-     */
-    public function report(string $path, LoggerInterface $logger): self
-    {
-        $this->builder->setReportUri($path);
-        $this->reportPath = $path;
-        $this->reportLogger = $logger;
-
-        return $this;
-    }
-
-    /**
      * Process a request and return a response.
      */
     public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
     {
-        if ($this->logReport($request)) {
-            return Factory::createResponse();
-        }
-
         $response = $handler->handle($request);
 
         $this->builder->compile();
@@ -84,30 +56,5 @@ class Csp implements MiddlewareInterface
             'object-src' => ['self' => true],
             'frame-ancestors' => ['self' => true],
         ]);
-    }
-
-    /**
-     * Handle the csp-report
-     * Returns true if the request is a report, false otherwise
-     */
-    private function logReport(ServerRequestInterface $request): bool
-    {
-        if ($request->getMethod() !== 'POST') {
-            return false;
-        }
-
-        if (empty($this->reportLogger) || $request->getUri()->getPath() !== $this->reportPath) {
-            return false;
-        }
-
-        $data = (array) $request->getParsedBody();
-
-        if (!isset($data['csp-report'])) {
-            return false;
-        }
-
-        $this->reportLogger->error('CSP error', $data);
-
-        return true;
     }
 }